delivery_bot/requirements.txt
psutil
pyautogui
python-dotenv
psycopg2-binary
pyperclip
